Add state change triggers to web_app.py
